[#10956] fix(catalog-glue): Move AWS Glue integration tests to integration/test path
